made changes based off comments
