+++ client/src/App.tsx
@@ -4,13 +4,11 @@ import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import Login from './pages/Login/Login';
 import Signup from './pages/SignUp/SignUp';
 import Dashboard from './pages/Dashboard/Dashboard';
-import { AuthProvider } from './context/useAuthContext';
 import { SocketProvider } from './context/useSocketContext';
 import { SnackBarProvider } from './context/useSnackbarContext';
 import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute';
 import './App.css';
 import Calendar from './pages/Calendar/Calendar';
-import { BoardProvider } from './helpers/APICalls/board';
 import { AuthBoardProvider } from './context/useAuthBoardContext';
 
 function App(): JSX.Element {
