+++ client/src/interface/IAuthBoardContext.ts
@@ -1,0 +1,8 @@
+import { AuthBoardApiDataSuccess } from './AuthBoardApiData';
+import { UserBoard } from './UserBoard';
+
+export default interface IAuthBoardContext {
+  loggedInUserBoard: UserBoard | null | undefined;
+  updateLoginContext: (data: AuthBoardApiDataSuccess) => void;
+  logout: () => void;
+}
